requirements.txt
aiobotocore==1.2.2
aiohttp==3.7.4.post0
aioitertools==0.7.1
amazon-mws-client==0.13.4
appdirs==1.4.4
args==0.1.0
arrow==0.17.0
asgiref==3.3.4
astroid==2.5.2
async-timeout==3.0.1
asyncssh==2.5.0
attrs==20.3.0
aws-google-auth==0.0.37
Babel==2.9.0
backcall==0.2.0
beautifulsoup4==4.9.3
binaryornot==0.4.4
black==21.4b0
boto3==1.16.52
botocore==1.19.52
bump2version==1.0.1
bumpversion==0.6.0
CacheControl==0.12.6
cached-property==1.5.2
cachy==0.3.0
cairocffi==1.2.0
CairoSVG==2.5.2
certifi==2021.5.30
cffi==1.14.5
cfgv==3.2.0
channel-integration-utils==0.5.1
chardet==4.0.0
cleo==0.8.1
click==7.1.2
client-core==9.1.0
clikit==0.6.2
clint==0.5.1
colorama==0.4.4
configparser==5.0.2
cookiecutter==1.7.3
correios==6.4.8
coverage==5.5
crashtest==0.3.1
crayons==0.4.0
cryptography==3.4.7
cssselect2==0.4.1
defusedxml==0.7.1
distlib==0.3.1
Django==3.2
django-filter==2.4.0
djangorestframework==3.12.4
dnspython==2.1.0
docutils==0.15.2
elastic-apm==6.1.1
email-validator==1.1.2
factory-boy==3.2.0
Faker==8.0.0
filelock==3.0.12
flake8==3.9.1
freezegun==1.1.0
fuzzywuzzy==0.18.0
gtin-validator==1.0.3
heroku3==4.2.3
html5lib==1.1
identify==2.2.4
idna==3.2
iniconfig==1.1.1
ipdb==0.13.7
ipython==7.22.0
ipython-genutils==0.2.0
isodate==0.6.0
isort==5.8.0
jadlog==0.27.3
jedi==0.18.0
jeepney==0.6.0
Jinja2==2.11.3
jinja2-time==0.2.0
jmespath==0.10.0
jplephem==2.15
json-encoder==0.4.4
keyring==21.8.0
keyrings.alt==4.0.2
lazy-object-proxy==1.6.0
lockfile==0.12.2
loggi==5.5.3
lunardate==0.2.0
lxml==4.6.3
Markdown==3.3.4
MarkupSafe==1.1.1
mccabe==0.6.1
msgpack==1.0.2
multidict==5.1.0
mypy-extensions==0.4.3
newrelic==5.22.1.152
nodeenv==1.6.0
olist-aws==6.1.3
olist-clients==41.11.0
olist-commons==7.1.0
olist-correios==6.5.0
olist-payday==1.5.1
olist-product-validation==7.10.0
packaging==20.9
parso==0.8.2
pastel==0.2.1
pathspec==0.8.1
pbr==5.5.1
pendulum==2.1.2
pep517==0.10.0
pexpect==4.8.0
phonenumbers==8.12.21
pickleshare==0.7.5
Pillow==8.2.0
Pint==0.17
pip-tools==6.1.0
pkginfo==1.7.0
pluggy==0.13.1
poetry==1.1.5
poetry-core==1.0.2
poyo==0.5.0
pre-commit==2.12.1
prettyconf==2.2.1
prompt-toolkit==3.0.18
psycopg2-binary==2.8.6
ptyprocess==0.7.0
py==1.10.0
pyCalverter==1.6.1
pycodestyle==2.7.0
pycparser==2.20
pycpfcnpj==1.5.1
pydantic==1.8.1
pyflakes==2.3.1
Pygments==2.8.1
pylev==1.3.0
pylint==2.7.4
pylint-django==2.4.3
pylint-plugin-utils==0.6
pyluach==1.2.1
pyOpenSSL==20.0.1
pyparsing==2.4.7
Pyphen==0.10.0
pysimplemodel==2.4.3
pytest==6.2.3
pytest-cov==2.11.1
pytest-deadfixtures==2.2.1
pytest-freezegun==0.4.2
python-dateutil==2.8.2
python-json-logger==2.0.1
python-Levenshtein==0.12.2
python-slugify==5.0.2
python-status==1.0.1
pytz==2021.1
pytzdata==2020.1
PyYAML==5.4.1
redis==3.5.3
regex==2021.4.4
reportlab==3.5.66
requests==2.26.0
requests-aws4auth==1.0.1
requests-file==1.5.1
requests-toolbelt==0.9.1
responses==0.13.2
retry==0.9.2
s3transfer==0.3.6
SecretStorage==3.3.1
selenium==3.141.0
sgp4==2.18
shellingham==1.4.0
simplejson==3.17.2
singledispatch==3.6.1
six==1.15.0
skyfield==1.38
skyfield-data==3.0.0
soupsieve==2.2.1
speaklater==1.3
sqlparse==0.4.1
stevedore==3.3.0
style==1.1.0
tabulate==0.8.9
tapioca-cnova==0.10.2
tapioca-colombo==0.3.1
tapioca-iugu==0.3.0
tapioca-mercadolivre==1.33.0
tapioca-mirakl==1.2.0
tapioca-skyhub==0.14.0
tapioca-wrapper==2.0.1
template-to-pdf==0.1.0
tenacity==7.0.0
text-unidecode==1.3
tinycss2==1.1.0
toml==0.10.2
tomlkit==0.7.0
traitlets==5.0.5
typing-extensions==3.7.4.3
tzlocal==2.1
Unidecode==1.2.0
update==0.0.1
urllib3==1.26.6
vcrpy==4.1.1
virtualenv==20.4.3
virtualenv-clone==0.5.4
virtualenvwrapper==4.8.4
wcwidth==0.2.5
WeasyPrint==52.4
webdriver-manager==3.4.2
webencodings==0.5.1
workalendar==15.1.0
wrapt==1.12.1
xmltodict==0.12.0
yarl==1.6.3
zeep==4.0.0
